# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Nsangusa API
  version: 1.0.0
  description: Static language-neutral contract for implemented Spring MVC routes.
    Spring Security supplies the explicitly marked login and logout operations.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://api.example.invalid
  description: Placeholder; configure per environment
...
security:
- sessionCookie: []
- basicAuth: []
paths:
  "/api/v1/auth/csrf":
    get:
      tags:
      - Identity
      operationId: getCsrfToken
      summary: Issue the CSRF cookie and return its request-header value
      security: []
      responses:
        '200':
          description: CSRF token
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CsrfTokenResponse"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/auth/register":
    post:
      tags:
      - Identity
      operationId: register
      summary: Request an account requiring email verification
      description: >
        Returns the same accepted response for a new or already registered address, without disclosing an account
        identifier. A new registration sends a real verification email. Existing accounts are not modified; use
        the verification request operation to request another verification link.
      security: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityRegistrationRequest'
      responses:
        '202':
          description: Registration request accepted without disclosing account existence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phase4IdentityRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  "/api/v1/auth/verify-email":
    post:
      tags:
      - Identity
      operationId: verifyEmail
      summary: Explicitly consume a single-use email-verification token
      description: >
        Only POST consumes the token. Missing, malformed, expired, previously consumed or wrong-purpose tokens are
        rejected. Token consumption is serialized in PostgreSQL. The default verification lifetime is 30 minutes
        and is configurable. Newsletter account linking occurs only after email verification.
      security: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityTokenRequest'
      responses:
        '204':
          description: Email verified.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/auth/password-reset/request":
    post:
      tags:
      - Identity
      operationId: requestPasswordReset
      summary: Request a short-lived password-reset email
      description: >
        Returns an accepted response without disclosing account existence. Only enabled, non-deleted accounts
        receive mail. Issuing a new token invalidates earlier reset tokens for the account. The default lifetime
        is 15 minutes and is configurable.
      security: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityEmailRequest'
      responses:
        '202':
          description: Password-reset request accepted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/auth/password-reset/confirm":
    post:
      tags:
      - Identity
      operationId: confirmPasswordReset
      summary: Explicitly consume a reset token and replace the password
      description: >
        Hashes the new password with the configured Argon2id encoder, clears login lockout tracking and revokes
        existing sessions after commit. A persistent security stamp prevents older authenticated sessions or a
        concurrent login using the old password from remaining authorized if Redis cleanup fails. GET does not
        consume the token.
      security: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityPasswordResetConfirm'
      responses:
        '204':
          description: Password replaced and previous authentication invalidated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/auth/me":
    get:
      tags:
      - Identity
      operationId: getCurrentUser
      summary: Read the currently authenticated account and its profile version
      description: Account identity is derived from the authenticated principal, never a supplied actor identifier.
      responses:
        '200':
          description: Current account; the response must not be publicly cached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phase4IdentityUserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
    patch:
      tags:
      - Identity
      operationId: updateCurrentUser
      summary: Update profile fields using the version actually loaded
      description: >
        Requires the current account version, including preference-only updates. Display names are trimmed and
        must remain nonblank. A newsletter frequency update requires an active linked subscription and does not
        grant consent or confirm a pending subscription. The legacy instant input is normalized to immediate.
        Stale versions return 409; successful changes are audited.
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityProfileUpdate'
      responses:
        '200':
          description: Updated current profile with its new version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phase4IdentityUserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      tags:
      - Identity
      operationId: deleteCurrentUser
      summary: Confirm and anonymize the current account using its loaded version
      description: >
        Disables local sign-in, clears roles, anonymizes email and display name, removes verification/reset tokens
        and linked external identities, unsubscribes and unlinks the newsletter, and invalidates sessions.
        Required audit/consent records and published contributions may be retained; this is not a claim of
        cross-store or backup erasure. The last active administrator cannot delete their account. Deletion and
        administrator changes share a transaction lock.
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityAccountDeletion'
      responses:
        '204':
          description: Account anonymized and sign-in invalidated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/auth/me/export":
    get:
      tags:
      - Identity
      operationId: exportCurrentUserData
      summary: Download the current account's identity-data export
      description: >
        Includes the current profile, linked newsletter preference and external identity links. Does not contain
        password hashes or verification/reset tokens. Does not export published content, audit records or other
        modules' retained operational data. Export is audited.
      responses:
        '200':
          description: Scoped account-data export.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Content-Disposition:
              schema:
                type: string
                const: attachment; filename="account-data.json"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phase4IdentityAccountExport'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/auth/sessions":
    get:
      tags:
      - Identity
      operationId: listCurrentUserSessions
      summary: List sessions indexed to the current account
      description: >
        Uses the Redis indexed session repository and sorts most recently accessed first. Marks the calling
        browser session as current. Session-store absence is an error, not an empty success response.
      responses:
        '200':
          description: Sessions belonging to the authenticated account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Phase4IdentitySession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/auth/sessions/{sessionId}":
    delete:
      tags:
      - Identity
      operationId: revokeCurrentUserSession
      summary: Revoke one session owned by the current account
      description: >
        Rejects an unknown session or a session belonging to another account. Revoking the current session also
        invalidates the current servlet session. Successful revocation is audited.
      parameters:
      - $ref: '#/components/parameters/SessionId'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Selected session revoked.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /api/v1/auth/verification/request:
    post:
      tags:
      - Identity
      operationId: requestIdentityVerification
      summary: Request a fresh verification link without disclosing account existence
      description: >
        Sends a link only for an enabled, non-deleted, unverified account. Previously issued verification tokens
        for that account are invalidated. The response does not reveal whether an eligible account exists.
      security: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityEmailRequest'
      responses:
        '202':
          description: Verification request accepted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Redis-backed opaque Spring Session cookie
    basicAuth:
      type: http
      scheme: basic
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: Spring Security CSRF token matching the XSRF-TOKEN cookie
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Invalid path, query, header, or request-body input
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    Unauthorized:
      description: Authentication is required or failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Forbidden:
      description: Authentication succeeded but authorization or CSRF validation failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Conflict:
      description: The requested transition conflicts with current resource state
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    TooManyRequests:
      description: Per-client request rate exceeded
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Retry delay in seconds
    ServiceUnavailable:
      description: Rate limiting could not reach its Redis dependency
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    UnexpectedError:
      description: Unexpected API failure
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
    Identifier:
      type: string
      format: uuid
...
    CsrfTokenResponse:
      type: object
      additionalProperties: false
      required:
      - headerName
      - token
      properties:
        headerName:
          type: string
        token:
          type: string
    Problem:
      type: object
      additionalProperties: true
      required:
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    Phase4IdentityRole:
      type: string
      enum:
      - READER
      - MODERATOR
      - EDITOR
      - ADMINISTRATOR
    Phase4IdentityVersion:
      type: integer
      format: int64
      minimum: 0
    Phase4IdentityRegistrationRequest:
      type: object
      additionalProperties: false
      required:
      - email
      - password
      - displayName
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 320
        password:
          type: string
          format: password
          writeOnly: true
          minLength: 12
          maxLength: 128
          pattern: \S
        displayName:
          type: string
          minLength: 1
          maxLength: 100
          pattern: \S
    Phase4IdentityRegistrationResponse:
      type: object
      additionalProperties: false
      required:
      - status
      properties:
        status:
          type: string
          const: verification_required
    Phase4IdentityEmailRequest:
      type: object
      additionalProperties: false
      required:
      - email
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 320
    Phase4IdentityToken:
      type: string
      writeOnly: true
      minLength: 43
      maxLength: 43
      pattern: ^[A-Za-z0-9_-]{43}$
      description: Opaque 256-bit unpadded base64url token delivered by email; only its SHA-256 hash is stored.
    Phase4IdentityTokenRequest:
      type: object
      additionalProperties: false
      required:
      - token
      properties:
        token:
          $ref: '#/components/schemas/Phase4IdentityToken'
    Phase4IdentityPasswordResetConfirm:
      type: object
      additionalProperties: false
      required:
      - token
      - newPassword
      properties:
        token:
          $ref: '#/components/schemas/Phase4IdentityToken'
        newPassword:
          type: string
          format: password
          writeOnly: true
          minLength: 12
          maxLength: 128
          pattern: \S
    Phase4IdentityProfileUpdate:
      type: object
      additionalProperties: false
      required:
      - expectedVersion
      properties:
        displayName:
          type:
          - string
          - 'null'
          minLength: 1
          maxLength: 100
          pattern: \S
        newsletterFrequency:
          type:
          - string
          - 'null'
          enum:
          - instant
          - immediate
          - daily
          - weekly
          - null
          description: Null or omitted leaves frequency unchanged; instant is a legacy alias for immediate.
        expectedVersion:
          $ref: '#/components/schemas/Phase4IdentityVersion'
    Phase4IdentityAccountDeletion:
      type: object
      additionalProperties: false
      required:
      - confirmation
      - expectedVersion
      properties:
        confirmation:
          type: string
          const: DELETE
        expectedVersion:
          $ref: '#/components/schemas/Phase4IdentityVersion'
    Phase4IdentityNewsletterPreference:
      type: object
      additionalProperties: false
      required:
      - subscriptionId
      - status
      - frequency
      properties:
        subscriptionId:
          $ref: '#/components/schemas/Identifier'
        status:
          type: string
          enum:
          - pending
          - confirmed
          - unsubscribed
          - suppressed
        frequency:
          type: string
          description: Current stored newsletter frequency; profile writes use immediate, daily or weekly.
    Phase4IdentityUserProfile:
      type: object
      additionalProperties: false
      required:
      - id
      - email
      - displayName
      - roles
      - emailVerified
      - createdAt
      - lastLoginAt
      - newsletter
      - version
      properties:
        id:
          $ref: '#/components/schemas/Identifier'
        email:
          type: string
          format: email
          maxLength: 320
        displayName:
          type: string
          maxLength: 100
        roles:
          type: array
          uniqueItems: true
          maxItems: 4
          items:
            $ref: '#/components/schemas/Phase4IdentityRole'
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type:
          - string
          - 'null'
          format: date-time
        newsletter:
          oneOf:
          - $ref: '#/components/schemas/Phase4IdentityNewsletterPreference'
          - type: 'null'
        version:
          $ref: '#/components/schemas/Phase4IdentityVersion'
    Phase4IdentityExternalIdentity:
      type: object
      additionalProperties: false
      required:
      - issuer
      - subject
      - emailAtLink
      - linkedAt
      - lastLoginAt
      properties:
        issuer:
          type: string
          maxLength: 500
        subject:
          type: string
          maxLength: 500
        emailAtLink:
          type:
          - string
          - 'null'
          format: email
          maxLength: 320
        linkedAt:
          type: string
          format: date-time
        lastLoginAt:
          type:
          - string
          - 'null'
          format: date-time
    Phase4IdentityAccountExport:
      type: object
      additionalProperties: false
      required:
      - generatedAt
      - profile
      - externalIdentities
      properties:
        generatedAt:
          type: string
          format: date-time
        profile:
          $ref: '#/components/schemas/Phase4IdentityUserProfile'
        externalIdentities:
          type: array
          items:
            $ref: '#/components/schemas/Phase4IdentityExternalIdentity'
    Phase4IdentitySession:
      type: object
      additionalProperties: false
      required:
      - id
      - createdAt
      - lastAccessedAt
      - expiresAt
      - current
      properties:
        id:
          type: string
          description: Opaque session identifier, not an account identifier.
        createdAt:
          type: string
          format: date-time
        lastAccessedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        current:
          type: boolean
...